Split serve into provisionerapi and added create command
