+++ cmd/provisioner.go
@@ -15,5 +15,6 @@ func init() {
 	provisionerCmd.AddCommand(provisionerServeCmd)
 	provisionerCmd.AddCommand(provisionerListCmd)
 	provisionerCmd.AddCommand(provisionerDeleteCmd)
+	provisionerCmd.AddCommand(provisionerCreateCmd)
 	rootCmd.AddCommand(provisionerCmd)
 }
